requirements.txt
amqp==5.1.1
aniso8601==9.0.1
anyio==3.6.1
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.0.8
async-lru==2.0.4
async-timeout==4.0.2
attrs==23.2.0
Babel==2.14.0
backcall==0.2.0
beautifulsoup4==4.11.1
billiard==4.1.0
bleach==5.0.1
blinker==1.9.0
Brotli==1.0.9
camelot-py==0.11.0
celery==5.3.4
certifi==2022.6.15
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
configparser==3.5.3
cryptography==41.0.4
cycler==0.11.0
dash==2.5.1
dash-bootstrap-components==1.1.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.4
Django==4.2.14
entrypoints==0.4
et-xmlfile==1.1.0
executing==1.0.0
fastapi==0.79.0
fastjsonschema==2.16.1
filelock==3.7.1
Flask==3.1.0
fonttools==4.37.1
fqdn==1.5.1
ghostscript==0.7
greenlet==2.0.2
h11==0.13.0
httpcore==1.0.2
httptools==0.4.0
httpx==0.26.0
idna==3.3
importlib-metadata==4.11.4
ipykernel==6.15.2
ipython==8.5.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.18.1
Jinja2==3.1.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter-events==0.9.0
jupyter-lsp==2.2.2
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.12.5
jupyter_server_terminals==0.5.2
jupyterlab==4.1.0
jupyterlab-widgets==3.0.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
kiwisolver==1.4.4
kombu==5.3.2
line-profiler==3.5.1
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
memory-profiler==0.60.0
mistune==2.0.4
nbclient==0.6.8
nbconvert==7.16.0
nbformat==5.9.2
nest-asyncio==1.5.5
notebook_shim==0.2.3
numpy==1.22.4
opencv-python==4.8.0.76
openpyxl==3.1.2
overrides==7.7.0
packaging==21.3
# pandas==2.2.1
pandocfilters==1.5.0
parso==0.8.3
pdfminer.six==20221105
# pdftopng==0.2.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pipenv==2022.7.4
platformdirs==2.5.2
plotly==5.8.2
prometheus-client==0.14.1
prompt-toolkit==3.0.39
psutil==5.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pydantic==1.9.1
Pygments==2.13.0
pyparsing==3.0.9
pypdf==4.1.0
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.20.0
python-json-logger==2.0.7
pytz==2022.1
PyYAML==6.0
pyzmq==25.1.2
QtPy==2.2.0
redis==4.3.5
referencing==0.33.0
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.17.1
Send2Trash==1.8.2
simple-term-menu==1.4.1
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.2.post1
SQLAlchemy==2.0.21
sqlparse==0.4.2
stack-data==0.5.0
starlette==0.19.1
tabulate==0.8.9
tenacity==8.0.1
terminado==0.15.0
tinycss2==1.1.1
tomli==2.0.1
tornado==6.2
traitlets==5.14.1
types-python-dateutil==2.8.19.20240106
typing_extensions==4.3.0
tzdata==2023.3
uri-template==1.3.0
urllib3==2.2.0
uvicorn==0.18.2
uvloop==0.16.0
vine==5.0.0
virtualenv==20.15.1
virtualenv-clone==0.5.7
watchfiles==0.16.1
wcwidth==0.2.5
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
websockets==10.3
Werkzeug==3.1.3
widgetsnbextension==4.0.3
zipp==3.8.0